% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buildStockModel.R
\name{buildStockModel}
\alias{buildStockModel}
\title{Build Stock Model}
\usage{
buildStockModel(data, yearColumn = "Year", valueColumn = "Value",
  groupingColumns = c("AreaCode", "ItemCode"), cumulativeYears = 15,
  plot = FALSE)
}
\arguments{
\item{data}{A data.table containing the data on the historical stock
changes.}

\item{yearColumn}{The column name of data which contains the years.}

\item{valueColumn}{The column name of data which contains the values of the
historical stock changes.}

\item{groupingColumns}{The column name(s) of data which contain the
variables which should be used for grouping.  Usually this will be country
and commodity variables: we don't want to use all data when building the
model because a stock change of 1000 for one country and commodity is very
different than a change of 1000 for a different country/commodity.  This
argument allows the user to specify these "groups" and then each group is
scaled according to it's variance (i.e. all stock changes are divided by
the standard deviation).  Then, we estimate the model on all the scaled
data.  Note that the cumulative stock computation will not be correct
without the correct grouping columns.}

\item{cumulativeYears}{Numeric.  How many years should be used to compute
the historical cumulative stock changes?  Defaults to 15.}

\item{plot}{Logical.  Should a plot be generated which shows the observed
  data as well as the model fit?  The error bars represent the confidence
  interval of the mean, not a prediction interval (and thus we don't expect
  that most observations will fall inside this interval).}
}
\value{
A list with three elements

  \itemize{

  \item model An object of type lm to use for prediction

  \item yearColumn The passed yearColumn.

  \item valueColumn The passed valueColumn.

  \item groupingColumns The passed groupingColumns.

  \item cumulativeYears The passed cumulativeYears. }
}
\description{
This function takes a dataset of historical, observed stock changes and
estimates a simple, linear regression model using the cumulative previous
drawdowns/inputs as the independent variable and the current drawdown/input
as the dependent variable.  Normality seemed like a reasonable assumption
after some initial testing, and so we feel comfortable with a simple linear
regression.  Thus, the distribution of the stock change in a year t, given
the previous cumulative drawdowns/inputs, will be normal with some mean and
variance estimated from this model.
}

