% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseStockModel.R
\name{chooseStockModel}
\alias{chooseStockModel}
\title{Choose Stock Model}
\usage{
chooseStockModel(fileNames)
}
\arguments{
\item{fileNames}{A vector of file names.}
}
\value{
The file name which should be loaded to provide the model.
}
\description{
This function takes a vector of strings (of .RData objects that exist in
some directory) and chooses the one that corresponds to the model that
should be used.
}

