% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeOpeningStocksfromDelta.R
\name{computeOpeningStocksfromDelta}
\alias{computeOpeningStocksfromDelta}
\title{Compute Opening Stocks From Stock Variation}
\usage{
computeOpeningStocksfromDelta(year)
}
\arguments{
\item{year}{Numeric.}
}
\value{
The dataset with the opening stocks computed.
}
\description{
This function pulls the old total trade data from FAOSTAT1
}

