% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProductionData.R
\name{getProductionData}
\alias{getProductionData}
\title{Get Production Data}
\usage{
getProductionData(yearRange)
}
\arguments{
\item{yearRange}{Character vector with the range of the years.}
}
\value{
The dataset with the production data.
}
\description{
This function pulls the production data from the SWS
}

