% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStockData.R
\name{getStockData}
\alias{getStockData}
\title{Get Stocks Data}
\usage{
getStockData(measuredElement, yearRange)
}
\arguments{
\item{yearRange}{Character vector with the range of the years.}
}
\value{
The dataset with the stock variation.
}
\description{
This function pulls the stock variation data from the SWS
}

