% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTotalTradeData.R
\name{getTotalTradeData}
\alias{getTotalTradeData}
\title{Get Total Trade Data}
\usage{
getTotalTradeData(yearRange)
}
\arguments{
\item{yearRange}{Character vector with the range of the years.}
}
\value{
The dataset with the total trade data.
}
\description{
This function pulls the total trade data from the SWS
}

