% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTotalTradeDataFAOSTAT1.R
\name{getTotalTradeDataFAOSTAT1}
\alias{getTotalTradeDataFAOSTAT1}
\title{Get Total Trade Data From FAOSTAT1}
\usage{
getTotalTradeDataFAOSTAT1(yearRange)
}
\arguments{
\item{yearRange}{Character vector with the range of the years.}
}
\value{
The dataset with the total trade data.
}
\description{
This function pulls the old total trade data from FAOSTAT1
}

