% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictStockModel.R
\name{predictStockModel}
\alias{predictStockModel}
\title{Predict Stock Model}
\usage{
predictStockModel(model, newdata, estimateYear, warn = FALSE)
}
\arguments{
\item{model}{A list object containing the relevant model information. Rather
than constructing this directly, it is easier to just pass the result of a
call to buildStockModel.}

\item{newdata}{A data.table containing the data which should be used for 
prediction.  If not provided, prediction is performed on the data used to 
fit the model.}

\item{estimateYear}{The year for which stock estimates should be generated. 
Must be a single value.}

\item{warn}{Logical, defaults to FALSE.  The variability must be estimated 
on data prior to the estimateYear.  If data after estimateYear exists, an 
error is thrown by default or a warning generated (if TRUE) and data after
estimateYear is deleted.}
}
\value{
A data.table object with the predictions.
}
\description{
This function takes a fitted model (from buildStockModel) and fits it to a 
provided dataset.  Essentially, this just requires calling predict.lm on the
newdata, but the complicated bit is that the cumulative stock changes must 
also be computed for estimation.
}

