% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveStockData.R
\name{saveStockData}
\alias{saveStockData}
\title{Save Stock Data}
\usage{
saveStockData(data, verbose = FALSE, stockCode = "5071")
}
\arguments{
\item{data}{The data.table containing the stock data, as would be passed to
SaveData.}

\item{verbose}{Logical.  Should messages be printed updating the user on
what is happening?}

\item{stockCode}{The element code for the stock commodity.}
}
\value{
No R objects are returned, as this functions purpose is solely to
  write to the database.
}
\description{
This function is essentially a SaveData call, but it does a few checks and
"clean-ups" to the stock data to prepare it for saving.
}

