% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getStockData.R
\name{getStockData}
\alias{getStockData}
\title{Get Stock Data}
\usage{
getStockData()
}
\value{
The data, stored in a data.table format.
}
\description{
This function pulls historical, measured data on stock changes.  Currently
it's just reading a .csv file, as that's all we have.
}

