% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predictStockModel.R
\name{predictStockModel}
\alias{predictStockModel}
\title{Predict Stock Model}
\usage{
predictStockModel(model, newdata)
}
\arguments{
\item{model}{A list object containing the relevant model information. Rather
than constructing this directly, it is easier to just pass the result of a
call to buildStockModel.}

\item{newdata}{A data.table containing the data which should be used for
  prediction.  If not provided, prediction is performed on the data used to
  fit the model.}
}
\value{
A data.table object with the predictions.
}
\description{
This function takes a fitted model (from buildStockModel) and fits it to a
provided dataset.  Essentially, this just requires calling predict.lm on the
newdata, but the complicated bit is that the cumulative stock changes must
also be computed for estimation.
}

